API/requirements.txt
blinker==1.7.0
boto3==1.34.84
botocore==1.34.84
click==8.1.7
Flask==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
pillow==10.3.0
python-dateutil==2.9.0.post0
s3transfer==0.10.1
six==1.16.0
urllib3==2.2.1
Werkzeug==3.0.2
